Working on refactoring for expansion of application
